requirements.txt
pytest==8.2.2
pytest-selenium==4.1.0
webdriver-manager==4.0.1
